Update codestyle in spigot-core module

+++ animatedarchitecture-spigot/spigot-core/src/main/java/nl/pim16aap2/animatedarchitecture/spigot/core/implementations/AudioPlayerSpigot.java
@@ -17,26 +17,35 @@ import java.util.function.DoubleUnaryOperator;
 /**
  * Represents an implementation of {@link IAudioPlayer} for the Spigot platform.
  */
-@Singleton
-public class AudioPlayerSpigot implements IAudioPlayer
+@Singleton public class AudioPlayerSpigot implements IAudioPlayer
 {
 
     private final IExecutor executor;
 
-    @Inject
-    public AudioPlayerSpigot(IExecutor executor)
+    @Inject public AudioPlayerSpigot(IExecutor executor)
     {
         this.executor = executor;
     }
 
-    @Override
-    public void playSound(
-        double x, double y, double z, IWorld world, String sound, float volume, float pitch, double range,
+    @Override public void playSound(
+        double x,
+        double y,
+        double z,
+        IWorld world,
+        String sound,
+        float volume,
+        float pitch,
+        double range,
         @Nullable DoubleUnaryOperator attenuationFunction)
     {
         playSound(
             new Location(Bukkit.getWorld(world.worldName()), x, y, z),
-            sound, volume, pitch, range, attenuationFunction);
+            sound,
+            volume,
+            pitch,
+            range,
+            attenuationFunction
+        );
     }
 
     private void playSound(
@@ -55,7 +64,12 @@ public class AudioPlayerSpigot implements IAudioPlayer
     }
 
     private void playSound(
-        Location loc, World world, String sound, float volume, float pitch, double range,
+        Location loc,
+        World world,
+        String sound,
+        float volume,
+        float pitch,
+        double range,
         @Nullable DoubleUnaryOperator attenuationFunction)
     {
         for (final Entity ent : world.getNearbyEntities(loc, range, range, range, e -> e instanceof Player))
@@ -64,7 +78,11 @@ public class AudioPlayerSpigot implements IAudioPlayer
     }
 
     private void playSound(
-        Player player, Location loc, String sound, float volume, float pitch,
+        Player player,
+        Location loc,
+        String sound,
+        float volume,
+        float pitch,
         @Nullable DoubleUnaryOperator attenuationFunction)
     {
         final double distance = player.getLocation().distance(loc);
@@ -73,7 +91,7 @@ public class AudioPlayerSpigot implements IAudioPlayer
 
         float attenuatedVolume = volume;
         if (attenuationFunction != null)
-            attenuatedVolume *= attenuationFunction.applyAsDouble(distance);
+            attenuatedVolume *= (float) attenuationFunction.applyAsDouble(distance);
 
         player.playSound(loc, sound, attenuatedVolume, pitch);
     }
